Extract input sender into composable module
- Create input-sender.js for buffered terminal input
- Uses requestAnimationFrame for batching
- Tries P2P first, falls back to WebSocket
- Remove ~11 lines from app.js (now 910 lines)
- All tests passing (666 pass)
